third_party/blink/web_tests/svg/dynamic-updates/SVGFESpecularLightingElement-inherit-lighting-color-css-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the inherited 'lighting-color' attribute of the SVGFESpecularLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.defaultView.getComputedStyle(specularElement).getPropertyValue('lighting-color') is "rgb(255, 0, 0)"
PASS document.defaultView.getComputedStyle(specularElement).getPropertyValue('lighting-color') is "rgb(173, 255, 47)"
PASS successfullyParsed is true

TEST COMPLETE
